expensetracker_backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
mysqlclient==2.2.0 